refactor(playground): split large files for agent-friendly codebase
Split monolithic files into small, single-responsibility modules:

- CustomComputerModal.tsx (674 lines) → AddSandboxModal, SourcePicker,
  PresetGrid, LocalDockerForm, CustomUrlForm, TestConnectionButton
- ChatInput.tsx (598 lines) → ChatInput, ComputerSelector,
  ModelSelector, MobileInputBar
- ChatContent.tsx → extract useComputerMapping, useSelectedComputer hooks
