+++ libs/typescript/playground/src/components/modals/index.ts
@@ -1,6 +1,6 @@
 // Barrel export for modals
 export { SettingsModal } from './SettingsModal';
-export { AddSandboxModal, CustomComputerModal } from './CustomComputerModal';
-export type { AddSandboxModalProps } from './CustomComputerModal';
+export { AddSandboxModal, CustomComputerModal } from './AddSandboxModal';
+export type { AddSandboxModalProps } from './AddSandboxModal';
 export { default as ExportTrajectoryModal } from './ExportTrajectoryModal';
 export { default as ReplayTrajectoryModal } from './ReplayTrajectoryModal';
