+++ libs/typescript/playground/src/hooks/index.ts
@@ -11,3 +11,5 @@ export {
 } from './usePlayground';
 
 export { useAgentRequest } from './useAgentRequest';
+export { useComputerMapping } from './useComputerMapping';
+export { useSelectedComputer } from './useSelectedComputer';
