+++ libs/typescript/playground/src/index.ts
@@ -42,6 +42,8 @@ export {
   useFindDefaultModel,
 } from './hooks/usePlayground';
 export { useAgentRequest } from './hooks/useAgentRequest';
+export { useComputerMapping } from './hooks/useComputerMapping';
+export { useSelectedComputer } from './hooks/useSelectedComputer';
 export { useDarkMode } from './hooks/useDarkMode';
 
 // Primitive Components
@@ -83,8 +85,8 @@ export { PlaygroundContent } from './components/composed/PlaygroundContent';
 
 // Modals
 export { SettingsModal } from './components/modals/SettingsModal';
-export { AddSandboxModal, CustomComputerModal } from './components/modals/CustomComputerModal';
-export type { AddSandboxModalProps } from './components/modals/CustomComputerModal';
+export { AddSandboxModal, CustomComputerModal } from './components/modals/AddSandboxModal';
+export type { AddSandboxModalProps } from './components/modals/AddSandboxModal';
 export { ExportTrajectoryModal } from './components/modals';
 export { ReplayTrajectoryModal } from './components/modals';
